apps-microservices/crud-milvus-service/requirements.txt
fastapi[all]
uvicorn
python-dotenv
pymilvus
pytest